+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 ## Change log
 ### 4.7.7 (9/14/2016)
 * http: throw methodNotAllowedException for unknown http method
+* session: ignore all errors on decoding redis session value
 
 ### 4.7.6 (9/13/2016)
 * session: change session data encoding to JSON